refactor: consolidate github subscription ownership
