+++ src/test-kernel/agent/toolUse/GitHubSubscriptionProgressEvents.vitest.ts
@@ -8,9 +8,6 @@ import {
 } from '@agent/runtime/AgentRuntimeHost';
 import { createRunContext, withRunContext } from '@agent/runtime/RunContext';
 
-// Local imports - event bus
-import { bus } from '@eventBus/ProgressEventBus';
-
 // Local imports - tools
 import { emitGitHubSubscriptionChanged } from '@tools/github/subscriptionEventEmitter';
 
@@ -18,14 +15,10 @@ import { emitGitHubSubscriptionChanged } from '@tools/github/subscriptionEventEm
 import { createRecordingHost } from '../progressTestUtils';
 
 describe('GitHub subscription progress events', () => {
-  it('publishes binding changes to the bus and scoped runtime host', () => {
+  it('publishes binding changes through the scoped runtime host', () => {
     const active = createRecordingHost();
     const fallback = createRecordingHost();
     const previousDefault = getDefaultAgentRuntimeHost();
-    const busEvents: unknown[] = [];
-    const off = bus.on('repoSubscriptionBindingsChanged', (payload) => {
-      busEvents.push(payload);
-    });
     setDefaultAgentRuntimeHost(fallback.host);
 
     try {
@@ -36,24 +29,18 @@ describe('GitHub subscription progress events', () => {
         );
       });
     } finally {
-      off();
       setDefaultAgentRuntimeHost(previousDefault);
     }
 
-    expect(busEvents).toEqual([undefined]);
     expect(active.events).toEqual([
       { event: 'repoSubscriptionBindingsChanged', payload: undefined },
     ]);
     expect(fallback.events).toEqual([]);
   });
 
-  it('does not duplicate binding changes through the default runtime host', () => {
+  it('uses the default runtime host outside a scoped run context', () => {
     const fallback = createRecordingHost();
     const previousDefault = getDefaultAgentRuntimeHost();
-    const busEvents: unknown[] = [];
-    const off = bus.on('issueSubscriptionBindingsChanged', (payload) => {
-      busEvents.push(payload);
-    });
     setDefaultAgentRuntimeHost(fallback.host);
 
     try {
@@ -62,11 +49,11 @@ describe('GitHub subscription progress events', () => {
         undefined,
       );
     } finally {
-      off();
       setDefaultAgentRuntimeHost(previousDefault);
     }
 
-    expect(busEvents).toEqual([undefined]);
-    expect(fallback.events).toEqual([]);
+    expect(fallback.events).toEqual([
+      { event: 'issueSubscriptionBindingsChanged', payload: undefined },
+    ]);
   });
 });
